Move LoadData to script_traits.
