+++ components/compositing/compositor.rs
@@ -19,14 +19,14 @@ use image::{DynamicImage, ImageFormat, RgbImage};
 use ipc_channel::ipc::{self, IpcSender, IpcSharedMemory};
 use ipc_channel::router::ROUTER;
 use msg::constellation_msg::{Image, PixelFormat, Key, KeyModifiers, KeyState};
-use msg::constellation_msg::{LoadData, TraversalDirection, PipelineId};
-use msg::constellation_msg::{PipelineIndex, PipelineNamespaceId, WindowSizeType};
+use msg::constellation_msg::{PipelineId, PipelineIndex, PipelineNamespaceId};
+use msg::constellation_msg::{TraversalDirection, WindowSizeType};
 use profile_traits::mem::{self, Reporter, ReporterRequest};
 use profile_traits::time::{self, ProfilerCategory, profile};
 use script_traits::{AnimationState, AnimationTickType, ConstellationControlMsg};
-use script_traits::{ConstellationMsg, LayoutControlMsg, MouseButton, MouseEventType};
-use script_traits::{StackingContextScrollState, TouchpadPressurePhase, TouchEventType};
-use script_traits::{TouchId, WindowSizeData};
+use script_traits::{ConstellationMsg, LayoutControlMsg, LoadData, MouseButton};
+use script_traits::{MouseEventType, StackingContextScrollState};
+use script_traits::{TouchpadPressurePhase, TouchEventType, TouchId, WindowSizeData};
 use script_traits::CompositorEvent::{self, MouseMoveEvent, MouseButtonEvent, TouchEvent, TouchpadPressureEvent};
 use std::collections::HashMap;
 use std::fs::File;
