+++ components/msg/Cargo.toml
@@ -14,13 +14,10 @@ bitflags = "0.7"
 cssparser = {version = "0.7", features = ["heap_size", "serde-serialization"]}
 heapsize = "0.3.0"
 heapsize_plugin = "0.1.2"
-hyper = "0.9.9"
-hyper_serde = "0.1.4"
 ipc-channel = "0.5"
 plugins = {path = "../plugins"}
 serde = "0.8"
 serde_derive = "0.8"
-url = {version = "1.2", features = ["heap_size", "serde"]}
 
 [dependencies.webrender_traits]
 git = "https://github.com/servo/webrender"
