+++ components/script/dom/htmlformelement.rs
@@ -42,9 +42,10 @@ use encoding::all::UTF_8;
 use encoding::label::encoding_from_whatwg_label;
 use hyper::header::{Charset, ContentDisposition, ContentType, DispositionParam, DispositionType};
 use hyper::method::Method;
-use msg::constellation_msg::{LoadData, PipelineId};
+use msg::constellation_msg::PipelineId;
 use rand::random;
 use script_thread::{MainThreadScriptMsg, Runnable};
+use script_traits::LoadData;
 use std::borrow::ToOwned;
 use std::cell::Cell;
 use std::sync::mpsc::Sender;
