+++ components/webdriver_server/lib.rs
@@ -33,12 +33,12 @@ use hyper::method::Method::{self, Post};
 use image::{DynamicImage, ImageFormat, RgbImage};
 use ipc_channel::ipc::{self, IpcReceiver, IpcSender};
 use keys::keycodes_to_keys;
-use msg::constellation_msg::{FrameId, LoadData, PipelineId};
+use msg::constellation_msg::{FrameId, PipelineId};
 use msg::constellation_msg::{PixelFormat, TraversalDirection};
 use regex::Captures;
 use rustc_serialize::base64::{CharacterSet, Config, Newline, ToBase64};
 use rustc_serialize::json::{Json, ToJson};
-use script_traits::{ConstellationMsg, WebDriverCommandMsg};
+use script_traits::{ConstellationMsg, LoadData, WebDriverCommandMsg};
 use script_traits::webdriver_msg::{LoadStatus, WebDriverCookieError, WebDriverFrameId};
 use script_traits::webdriver_msg::{WebDriverJSError, WebDriverJSResult, WebDriverScriptCommand};
 use std::borrow::ToOwned;
